clean

+++ src/assets/assets.service.ts
@@ -1,6 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  HttpCode,
+  HttpException,
+  HttpStatus,
+  Injectable,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Op } from 'sequelize';
+import { AssetPack } from 'src/asset-packs/entities/asset-pack.entity';
 import { PaginateQuery } from 'src/common/paginate/decorator';
 import { paginate } from 'src/common/paginate/paginate';
 import { CreateAssetDto } from './dto/create-asset.dto';
@@ -12,10 +18,21 @@ export class AssetsService {
   constructor(
     @InjectModel(Asset)
     private assetModel: typeof Asset,
+    @InjectModel(AssetPack)
+    private assetPack: typeof AssetPack,
   ) {}
 
-  create(ownerId: string, createAssetDto: CreateAssetDto) {
-    return this.assetModel.create({ ...createAssetDto, owner: ownerId });
+  async create(owner: string, createAssetDto: CreateAssetDto) {
+    const assetPack = await this.assetModel.findOne({
+      where: { id: createAssetDto.pack_id, owner },
+    });
+    if (!assetPack) {
+      throw new HttpException(
+        `asset pack ${createAssetDto.pack_id} not found`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    return this.assetModel.create({ ...createAssetDto, owner });
   }
 
   findAll(query: PaginateQuery) {
@@ -25,7 +42,8 @@ export class AssetsService {
       defaultSortBy: [['id', 'DESC']],
       filterableColumns: {
         id: [],
-        owner: [Op.in],
+        owner: [],
+        pack_id: [],
       },
     });
   }
@@ -34,11 +52,27 @@ export class AssetsService {
     return this.assetModel.findByPk(id);
   }
 
-  update(id: string, updateAssetDto: UpdateAssetDto) {
-    return this.assetModel.update(updateAssetDto, { where: { id } });
+  async update(owner: string, id: string, updateAssetDto: UpdateAssetDto) {
+    const asset = await this.assetModel.findOne({ where: { id, owner } });
+    if (!asset) {
+      throw new HttpException(`asset ${id} not found`, HttpStatus.NOT_FOUND);
+    }
+
+    if (updateAssetDto.pack_id && asset.pack_id != updateAssetDto.pack_id) {
+      const assetPack = await this.assetModel.findOne({ where: { id, owner } });
+      if (!assetPack) {
+        throw new HttpException(
+          `asset pack ${updateAssetDto.pack_id} not found`,
+          HttpStatus.NOT_FOUND,
+        );
+      }
+    }
+    asset.setAttributes(updateAssetDto);
+    await asset.save();
+    return asset;
   }
 
-  remove(id: string) {
-    return this.assetModel.destroy({ where: { id } });
+  remove(owner: string, id: string) {
+    return this.assetModel.destroy({ where: { id, owner } });
   }
 }
