+++ src/items/items.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Op } from 'sequelize';
+import { Collection } from 'src/collections/entities/collection.entity';
 import { PaginateQuery } from 'src/common/paginate/decorator';
 import { paginate } from 'src/common/paginate/paginate';
 import { CreateItemDto } from './dto/create-item.dto';
@@ -12,6 +13,8 @@ export class ItemsService {
   constructor(
     @InjectModel(Item)
     private itemModel: typeof Item,
+    @InjectModel(Collection)
+    private collectionModel: typeof Collection,
   ) {}
 
   create(authId: string, createItemDto: CreateItemDto) {
@@ -25,20 +28,36 @@ export class ItemsService {
       defaultSortBy: [['id', 'DESC']],
       filterableColumns: {
         id: [],
-        eth_address: [Op.in],
+        owner: [],
+        collection_id: [],
       },
     });
   }
 
-  findOne(id: number) {
+  findOne(id: string) {
     return this.itemModel.findByPk(id);
   }
 
-  update(id: number, updateItemDto: UpdateItemDto) {
-    return this.itemModel.update(updateItemDto, { where: { id: id } });
+  async update(owner: string, id: string, updateItemDto: UpdateItemDto) {
+    const item = await this.itemModel.findOne({ where: { id, owner } });
+    if (!item) {
+      throw new HttpException(`item ${id} not found`, HttpStatus.NOT_FOUND);
+    }
+    const collectionId = updateItemDto.collection_id || item.collection_id;
+    const collection = await this.collectionModel.findOne({
+      where: { id: collectionId, owner },
+    });
+
+    if (!collection)
+      throw new HttpException(
+        `collection ${collectionId} not found`,
+        HttpStatus.NOT_FOUND,
+      );
+
+    return this.itemModel.update(updateItemDto, { where: { id, owner } });
   }
 
-  remove(id: number) {
-    return this.itemModel.destroy({ where: { id } });
+  remove(owner: string, id: string) {
+    return this.itemModel.destroy({ where: { id, owner } });
   }
 }
